# apps/web/openapi.yaml
openapi: 3.0.0
info:
  title: GolfFox API
  description: API completa do sistema GolfFox - Gestão de transporte de funcionários
  version: 1.0.0
  contact:
    name: GolfFox Support
    email: support@golffox.com

servers:
  - url: http://localhost:3000/api
    description: Servidor de desenvolvimento
  - url: https://api.golffox.com/api
    description: Servidor de produção

tags:
  - name: Admin
    description: Endpoints administrativos
  - name: Operator
    description: Endpoints do operador
  - name: Carrier
    description: Endpoints da transportadora
  - name: Auth
    description: Autenticação
  - name: Health
    description: Health checks

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Verifica o status da aplicação e conexão com Supabase
      responses:
        '200':
          description: Aplicação saudável
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded, unhealthy]
                  ok:
                    type: boolean
                  checks:
                    type: object
                  timestamp:
                    type: string
                    format: date-time
                  duration:
                    type: string
        '500':
          description: Aplicação com problemas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /admin/companies:
    get:
      tags:
        - Admin
      summary: Listar empresas
      description: Lista todas as empresas com paginação e filtros
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de empresas
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  count:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  totalPages:
                    type: integer
                  hasNext:
                    type: boolean
                  hasPrev:
                    type: boolean
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro no servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Admin
      summary: Criar empresa
      description: Cria uma nova empresa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompany'
      responses:
        '201':
          description: Empresa criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Company'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /admin/companies/{companyId}:
    get:
      tags:
        - Admin
      summary: Buscar empresa por ID
      security:
        - bearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Empresa encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Company'
        '404':
          description: Empresa não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Admin
      summary: Atualizar empresa
      security:
        - bearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompany'
      responses:
        '200':
          description: Empresa atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Company'

    delete:
      tags:
        - Admin
      summary: Deletar empresa
      security:
        - bearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Empresa deletada
        '404':
          description: Empresa não encontrada

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Company:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        cnpj:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateCompany:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        cnpj:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email

    UpdateCompany:
      type: object
      properties:
        name:
          type: string
        cnpj:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        is_active:
          type: boolean

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string